refactor relational table ut
